move libc to the 'server' directory.
Finding the hash of this libc is part of the challenge, and its intended
that players find this file on their own. Due to the possibility of the
file falling off the internet, it is hosted in this repo. If
playtesting, you shouldn't obtain this libc until you have 'earned' it.